add render title #JC19-359
