+++ apps/vanilla/src/pages/detail/index.html
@@ -8,7 +8,6 @@
     <title>Detail</title>
   </head>
   <body class="default-layout">
-    <h1 class="title">Detail page</h1>
     <div class="anime-detail">
       <div class="anime-detail__media"></div>
       <div class="anime-detail__content"></div>
